Add files via upload
This is an application that helps u solve different math problems and formulas by submitting values. You can choose a formula in the home application and then it will take you to another window which will ask the user for the values. From there you can choose to solve or clear the entries.